TesteHD

+++ routes/web.php
@@ -18,18 +18,19 @@ Route::get('/register', function () {
 })->name('register');
 
 Route::prefix('user')->group(function () {
-    Route::get('/', ['as' => 'user.index', 'uses' => "App\Http\Controllers\UserController@index"]);
-    Route::get('/show/{id}', ['as' => 'user.show', 'uses' => "App\Http\Controllers\UserController@show"]);
-    Route::get('/create', ['as' => 'user.create', 'uses' => "App\Http\Controllers\UserController@create"]);
-    Route::post('/store', ['as' => 'user.store', 'uses' => "App\Http\Controllers\UserController@store"]);
-    Route::put('/edit/{id}', ['as' => 'user.edit', 'uses' => "App\Http\Controllers\UserController@edit"]);   
-    Route::put('/update/{id}', ['as' => 'user.update', 'uses' => "App\Http\Controllers\UserController@update"]);
-    Route::delete('/delete/{id}', ['as' => 'user.delete', 'uses' => "App\Http\Controllers\UserController@delete"]);
+    Route::get('/', ['as' => 'user.index', 'uses' => "App\Http\Controllers\admin\UserController@index"]);
+    Route::get('/all', ['as' => 'user.all', 'uses' => "App\Http\Controllers\admin\UserController@all"]);
+    Route::get('/show/{id}', ['as' => 'user.show', 'uses' => "App\Http\Controllers\admin\UserController@show"]);
+    Route::get('/create', ['as' => 'user.create', 'uses' => "App\Http\Controllers\admin\UserController@create"]);
+    Route::post('/store', ['as' => 'user.store', 'uses' => "App\Http\Controllers\admin\UserController@store"]);
+    Route::put('/edit/{id}', ['as' => 'user.edit', 'uses' => "App\Http\Controllers\admin\UserController@edit"]);   
+    Route::put('/update/{id}', ['as' => 'user.update', 'uses' => "App\Http\Controllers\admin\UserController@update"]);
+    Route::delete('/delete/{id}', ['as' => 'user.delete', 'uses' => "App\Http\Controllers\admin\UserController@delete"]);
 });
 
 Route::prefix('auth')->group(function () {
-    Route::post('/login', ['as' => 'auth.login', 'uses' => "App\Http\Controllers\AuthController@login"]);
-    Route::post('/logout', ['as' => 'auth.logout', 'uses' => "App\Http\Controllers\AuthController@logout"]);    
+    Route::post('/login', ['as' => 'auth.login', 'uses' => "App\Http\Controllers\Auth\AuthController@login"]);
+    Route::post('/logout', ['as' => 'auth.logout', 'uses' => "App\Http\Controllers\Auth\AuthController@logout"]);    
 });
 
 Route::prefix('exemplo')->group(function () {
